honor the tree-invariant in ShardedMultiGenerationBlobAccess
when a tree root is uploaded, an upstream of the whole tree is forced
soon after to guarantee that the whole tree is contained within one
generation.

to do not break the tree-invariant, a controller is spawned to
coordinate the rotations. when a shard needs to rotate, it sets a
flag. As soon as the controller detects it, it will trigger a
synchronous rotation of all the shards. this is fundamental, since the
blobs recursively referenced within a tree can be stored on different
shards.